refactor: combined api request extension files
Create new extension package `ApiExtension` and add ApiRequest files into it.
