+++ src/main/java/com/labs/restservice/triangle/TriangleController.java
@@ -2,7 +2,7 @@ package com.labs.restservice.triangle;
 
 import com.labs.restservice.calculations.CalculationResults;
 import com.labs.restservice.calculations.CalculationService;
-import com.labs.restservice.exception.ApiRequestException;
+import com.labs.restservice.exception.ApiException.ApiRequestException;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
